Refactor ModResolver
