+++ src/main/java/net/fabricmc/loader/impl/discovery/ModResolutionException.java
@@ -22,6 +22,10 @@ public class ModResolutionException extends Exception {
 		super(s);
 	}
 
+	public ModResolutionException(String format, Object... args) {
+		super(String.format(format, args));
+	}
+
 	public ModResolutionException(Throwable t) {
 		super(t);
 	}
