+++ src/main/java/net/fabricmc/loader/impl/discovery/ModResolver.java
@@ -16,16 +16,10 @@
 
 package net.fabricmc.loader.impl.discovery;
 
-import static com.google.common.jimfs.Feature.FILE_CHANNEL;
-import static com.google.common.jimfs.Feature.SECURE_DIRECTORY_STREAM;
-
-import java.io.File;
-import java.io.IOException;
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.nio.file.FileSystem;
-import java.nio.file.Files;
-import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -34,36 +28,14 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Queue;
 import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.RecursiveAction;
-import java.util.concurrent.TimeUnit;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import java.util.zip.ZipError;
-
-import com.google.common.jimfs.Configuration;
-import com.google.common.jimfs.Jimfs;
-import com.google.common.jimfs.PathType;
 
 import net.fabricmc.loader.api.Version;
 import net.fabricmc.loader.api.metadata.ModDependency;
 import net.fabricmc.loader.impl.FabricLoaderImpl;
-import net.fabricmc.loader.impl.game.GameProvider.BuiltinMod;
-import net.fabricmc.loader.impl.launch.FabricLauncherBase;
-import net.fabricmc.loader.impl.lib.gson.MalformedJsonException;
-import net.fabricmc.loader.impl.metadata.BuiltinModMetadata;
-import net.fabricmc.loader.impl.metadata.LoaderModMetadata;
 import net.fabricmc.loader.impl.metadata.ModMetadataParser;
-import net.fabricmc.loader.impl.metadata.NestedJarEntry;
-import net.fabricmc.loader.impl.metadata.ParseMetadataException;
-import net.fabricmc.loader.impl.util.FileSystemUtil;
 import net.fabricmc.loader.impl.util.UrlConversionException;
 import net.fabricmc.loader.impl.util.UrlUtil;
 import net.fabricmc.loader.impl.util.log.Log;
@@ -77,36 +49,30 @@ import net.fabricmc.loader.util.sat4j.specs.IVecInt;
 import net.fabricmc.loader.util.sat4j.specs.TimeoutException;
 
 public class ModResolver {
-	// nested JAR store
-	private static final FileSystem inMemoryFs = Jimfs.newFileSystem(
-			"nestedJarStore",
-			Configuration.builder(PathType.unix())
-			.setRoots("/")
-			.setWorkingDirectory("/")
-			.setAttributeViews("basic")
-			.setSupportedFeatures(SECURE_DIRECTORY_STREAM, FILE_CHANNEL)
-			.build()
-			);
-	private static final Map<String, List<Path>> inMemoryCache = new ConcurrentHashMap<>();
-	private static final Pattern MOD_ID_PATTERN = Pattern.compile("[a-z][a-z0-9-_]{1,63}");
-	private static final Object launcherSyncObject = new Object();
-
-	private final List<ModCandidateFinder> candidateFinders = new ArrayList<>();
-
 	public ModResolver() {
 	}
 
-	public void addCandidateFinder(ModCandidateFinder f) {
-		candidateFinders.add(f);
-	}
+	public Map<String, ModCandidate> resolve(Map<String, ModCandidateSet> candidatesById) throws ModResolutionException {
+		long time2 = System.currentTimeMillis();
+		Map<String, ModCandidate> result = findCompatibleSet(candidatesById);
 
-	private static IVecInt toVecInt(IntStream stream) {
-		return new VecInt(stream.toArray());
+		long time3 = System.currentTimeMillis();
+		Log.debug(LogCategory.RESOLUTION, "Mod resolution time: " + (time3 - time2) + "ms");
+
+		for (ModCandidate candidate : result.values()) {
+			if (candidate.getInfo().getSchemaVersion() < ModMetadataParser.LATEST_VERSION) {
+				Log.warn(LogCategory.METADATA, "Mod ID %s uses outdated schema version: %d < %d", candidate.getInfo().getId(), candidate.getInfo().getSchemaVersion(), ModMetadataParser.LATEST_VERSION);
+			}
+
+			candidate.getInfo().emitFormatWarnings();
+		}
+
+		return result;
 	}
 
 	// TODO: Find a way to sort versions of mods by suggestions and conflicts (not crucial, though)
 	@SuppressWarnings("unchecked")
-	public Map<String, ModCandidate> findCompatibleSet(Map<String, ModCandidateSet> modCandidateSetMap) throws ModResolutionException {
+	private Map<String, ModCandidate> findCompatibleSet(Map<String, ModCandidateSet> modCandidateSetMap) throws ModResolutionException {
 		// First, map all ModCandidateSets to Set<ModCandidate>s.
 		boolean isAdvanced = false;
 		Map<String, List<ModCandidate>> modCandidateMap = new HashMap<>();
@@ -169,7 +135,7 @@ public class ModResolver {
 							solver.addAtMost(versionVec, 1);
 						}
 					} catch (ContradictionException e) {
-						throw new ModResolutionException("Could not resolve valid mod collection (at: adding mod " + id + ")", e);
+						throw new ModResolutionException(String.format("Could not resolve valid mod collection (at: adding mod %s)", id), e);
 					}
 				}
 
@@ -196,7 +162,7 @@ public class ModResolver {
 							solver.addClause(new VecInt(clause));
 						} catch (ContradictionException e) {
 							//TODO: JiJ'd mods don't manage to throw this exception and instead fail silently
-							throw new ModResolutionException("Could not find required mod: " + mod.getInfo().getId() + " requires " + dep, e);
+							throw new ModResolutionException(String.format("Could not find required mod: %s requires %s", mod.getInfo().getId(), dep), e);
 						}
 					}
 
@@ -217,7 +183,7 @@ public class ModResolver {
 								solver.addClause(new VecInt(new int[] { -modClauseId, -m }));
 							}
 						} catch (ContradictionException e) {
-							throw new ModResolutionException("Found conflicting mods: " + mod.getInfo().getId() + " conflicts with " + dep, e);
+							throw new ModResolutionException(String.format("Found conflicting mods: %s conflicts with %s", mod.getInfo().getId(), dep), e);
 						}
 					}
 				}
@@ -243,7 +209,7 @@ public class ModResolver {
 
 					if (!satisfied) {
 						if (mandatoryMods.contains(mod)) {
-							throw new ModResolutionException("Could not resolve mod collection including mandatory mod '" + mod + "'");
+							throw new ModResolutionException("Could not resolve mod collection including mandatory mod '%s'", mod);
 						} else {
 							assumptions = assumptions.pop();
 						}
@@ -266,7 +232,7 @@ public class ModResolver {
 					ModCandidate candidate = intCandidateMap.get(i);
 
 					if (result.containsKey(candidate.getInfo().getId())) {
-						throw new ModResolutionException("Duplicate ID '" + candidate.getInfo().getId() + "' after solving - wrong constraints?");
+						throw new ModResolutionException("Duplicate ID '%s' after solving - wrong constraints?", candidate.getInfo().getId());
 					} else {
 						result.put(candidate.getInfo().getId(), candidate);
 					}
@@ -285,78 +251,81 @@ public class ModResolver {
 			}
 		}
 
-		StringBuilder errorsHard = new StringBuilder();
-		StringBuilder errorsSoft = new StringBuilder();
+		StringWriter errorsHardSw = new StringWriter();
+		StringWriter errorsSoftSw = new StringWriter();
 
-		if (!missingMods.isEmpty()) {
-			errorsHard.append("\n - Missing mods: ").append(String.join(", ", missingMods));
-		} else {
-			// verify result: dependencies
-			for (ModCandidate candidate : result.values()) {
-				for (ModDependency dependency : candidate.getInfo().getDepends()) {
-					addErrorToList(candidate, dependency, result, errorsHard, "requires", true);
-				}
+		try (PrintWriter errorsHardPw = new PrintWriter(errorsHardSw);
+				PrintWriter errorsSoftPw = new PrintWriter(errorsSoftSw)) {
+			if (!missingMods.isEmpty()) {
+				errorsHardPw.printf("\n - Missing mods: %s", String.join(", ", missingMods));
+			} else {
+				// verify result: dependencies
+				for (ModCandidate candidate : result.values()) {
+					for (ModDependency dependency : candidate.getInfo().getDepends()) {
+						addErrorToList(candidate, dependency, result, "requires", true, errorsHardPw);
+					}
 
-				for (ModDependency dependency : candidate.getInfo().getRecommends()) {
-					addErrorToList(candidate, dependency, result, errorsSoft, "recommends", true);
-				}
+					for (ModDependency dependency : candidate.getInfo().getRecommends()) {
+						addErrorToList(candidate, dependency, result, "recommends", true, errorsSoftPw);
+					}
 
-				for (ModDependency dependency : candidate.getInfo().getBreaks()) {
-					addErrorToList(candidate, dependency, result, errorsHard, "is incompatible with", false);
-				}
+					for (ModDependency dependency : candidate.getInfo().getBreaks()) {
+						addErrorToList(candidate, dependency, result, "is incompatible with", false, errorsHardPw);
+					}
 
-				for (ModDependency dependency : candidate.getInfo().getConflicts()) {
-					addErrorToList(candidate, dependency, result, errorsSoft, "conflicts with", false);
-				}
+					for (ModDependency dependency : candidate.getInfo().getConflicts()) {
+						addErrorToList(candidate, dependency, result, "conflicts with", false, errorsSoftPw);
+					}
 
-				Version version = candidate.getInfo().getVersion();
-				List<Version> suspiciousVersions = new ArrayList<>();
+					Version version = candidate.getInfo().getVersion();
+					List<Version> suspiciousVersions = new ArrayList<>();
 
-				for (ModCandidate other : modCandidateMap.get(candidate.getInfo().getId())) {
-					Version otherVersion = other.getInfo().getVersion();
+					for (ModCandidate other : modCandidateMap.get(candidate.getInfo().getId())) {
+						Version otherVersion = other.getInfo().getVersion();
 
-					if (version instanceof Comparable && otherVersion instanceof Comparable && !version.equals(otherVersion)) {
-						if (((Comparable<Version>) version).compareTo(otherVersion) == 0) {
-							suspiciousVersions.add(otherVersion);
+						if (version instanceof Comparable && otherVersion instanceof Comparable && !version.equals(otherVersion)) {
+							if (((Comparable<Version>) version).compareTo(otherVersion) == 0) {
+								suspiciousVersions.add(otherVersion);
+							}
 						}
 					}
-				}
 
-				if (!suspiciousVersions.isEmpty()) {
-					errorsSoft.append(String.format("\n - Conflicting versions found for %s: used %s, also found %s",
-							candidate.getInfo().getId(),
-							version.getFriendlyString(),
-							suspiciousVersions.stream().map(Version::getFriendlyString).collect(Collectors.joining(", "))));
+					if (!suspiciousVersions.isEmpty()) {
+						errorsSoftPw.printf("\n - Conflicting versions found for %s: used %s, also found %s",
+								candidate.getInfo().getId(),
+								version.getFriendlyString(),
+								suspiciousVersions.stream().map(Version::getFriendlyString).collect(Collectors.joining(", ")));
+					}
 				}
 			}
 		}
 
 		// print errors
-		String errHardStr = errorsHard.toString();
-		String errSoftStr = errorsSoft.toString();
 
-		if (!errSoftStr.isEmpty()) {
-			Log.warn(LogCategory.RESOLUTION, "Warnings were found! %s", errSoftStr);
-		}
-
-		if (!errHardStr.isEmpty()) {
-			throw new ModResolutionException("Errors were found!" + errHardStr + errSoftStr);
+		if (errorsHardSw.getBuffer().length() > 0) {
+			throw new ModResolutionException("Errors were found!%s%s", errorsHardSw, errorsSoftSw);
+		} else if (errorsSoftSw.getBuffer().length() > 0) {
+			Log.warn(LogCategory.RESOLUTION, "Warnings were found! %s", errorsSoftSw);
 		}
 
 		return result;
 	}
 
-	private void addErrorToList(ModCandidate candidate, ModDependency dependency, Map<String, ModCandidate> result, StringBuilder errors, String errorType, boolean cond) {
+	private static IVecInt toVecInt(IntStream stream) {
+		return new VecInt(stream.toArray());
+	}
+
+	private void addErrorToList(ModCandidate candidate, ModDependency dependency, Map<String, ModCandidate> result, String errorType, boolean cond, PrintWriter pw) {
 		String depModId = dependency.getModId();
 
 		List<String> errorList = new ArrayList<>();
 
-		if (!isModIdValid(depModId, errorList)) {
-			errors.append("\n - Mod ").append(getCandidateName(candidate)).append(" ").append(errorType).append(" ")
-			.append(depModId).append(", which has an invalid mod ID because:");
+		if (!ModDiscoverer.isModIdValid(depModId, errorList)) {
+			pw.printf("\n - Mod %s %s %s, which has an invalid mod ID because:",
+					getCandidateName(candidate), errorType, depModId);
 
 			for (String error : errorList) {
-				errors.append("\n\t - It ").append(error);
+				pw.printf("\n\t - It %s", error);
 			}
 
 			return;
@@ -369,7 +338,7 @@ public class ModResolver {
 			for (ModCandidate value : result.values()) {
 				if (value.getInfo().getProvides().contains(depModId)) {
 					if (FabricLoaderImpl.INSTANCE.isDevelopmentEnvironment()) {
-						Log.warn(LogCategory.METADATA, "Mod %s is using the provided alias %s in place of the real mod id %s.  Please use the mod id instead of a provided alias.",
+						Log.warn(LogCategory.METADATA, "Mod %s is using the provided alias %s in place of the real mod id %s. Please use the mod id instead of a provided alias.",
 								candidate.getInfo().getId(), depModId, value.getInfo().getId());
 					}
 
@@ -382,63 +351,60 @@ public class ModResolver {
 		boolean isPresent = depCandidate != null && dependency.matches(depCandidate.getInfo().getVersion());
 
 		if (isPresent != cond) {
-			errors.append("\n - Mod ").append(getCandidateName(candidate)).append(" ").append(errorType).append(" ")
-			.append(getDependencyVersionRequirements(dependency)).append(" of mod ")
-			.append(depCandidate == null ? depModId : getCandidateName(depCandidate)).append(", ");
+			pw.printf("\n - Mod %s %s %s of mod %s, ",
+					getCandidateName(candidate), errorType, getDependencyVersionRequirements(dependency),
+					(depCandidate == null ? depModId : getCandidateName(depCandidate)));
 
 			if (depCandidate == null) {
-				appendMissingDependencyError(errors, dependency);
+				appendMissingDependencyError(dependency, pw);
 			} else if (cond) {
-				appendUnsatisfiedDependencyError(errors, dependency, depCandidate);
+				appendUnsatisfiedDependencyError(dependency, depCandidate, pw);
 			} else if (errorType.contains("conf")) {
 				// CONFLICTS WITH
-				appendConflictError(errors, candidate, depCandidate);
+				appendConflictError(candidate, depCandidate, pw);
 			} else {
-				appendBreakingError(errors, candidate, depCandidate);
+				appendBreakingError(candidate, depCandidate, pw);
 			}
 
 			if (depCandidate != null) {
-				appendJiJInfo(errors, result, depCandidate);
+				appendJiJInfo(result, depCandidate, pw);
 			}
 		}
 	}
 
-	private void appendMissingDependencyError(StringBuilder errors, ModDependency dependency) {
-		errors.append("which is missing!");
-		errors.append("\n\t - You must install ").append(getDependencyVersionRequirements(dependency)).append(" of ")
-		.append(dependency.getModId()).append(".");
+	private void appendMissingDependencyError(ModDependency dependency, PrintWriter pw) {
+		pw.printf("which is missing!\n\t - You must install %s of %s.",
+				getDependencyVersionRequirements(dependency), dependency.getModId());
 	}
 
-	private void appendUnsatisfiedDependencyError(StringBuilder errors, ModDependency dependency, ModCandidate depCandidate) {
-		errors.append("but a non-matching version is present: ").append(getCandidateFriendlyVersion(depCandidate)).append("!");
-		errors.append("\n\t - You must install ").append(getDependencyVersionRequirements(dependency)).append(" of ")
-		.append(getCandidateName(depCandidate)).append(".");
+	private void appendUnsatisfiedDependencyError(ModDependency dependency, ModCandidate depCandidate, PrintWriter pw) {
+		pw.printf("but a non-matching version is present: %s!\n\t - You must install %s of %s.",
+				getCandidateFriendlyVersion(depCandidate), getDependencyVersionRequirements(dependency), getCandidateName(depCandidate));
 	}
 
-	private void appendConflictError(StringBuilder errors, ModCandidate candidate, ModCandidate depCandidate) {
+	private void appendConflictError(ModCandidate candidate, ModCandidate depCandidate, PrintWriter pw) {
 		final String depCandidateVer = getCandidateFriendlyVersion(depCandidate);
-		errors.append("but a matching version is present: ").append(depCandidateVer).append("!");
-		errors.append("\n\t - While this won't prevent you from starting the game,");
-		errors.append(" the developer(s) of ").append(getCandidateName(candidate));
-		errors.append(" have found that version ").append(depCandidateVer).append(" of ").append(getCandidateName(depCandidate));
-		errors.append(" conflicts with their mod.");
-		errors.append("\n\t - It is heavily recommended to remove one of the mods.");
+
+		pw.printf("but a matching version is present: %s!\n"
+				+ "\t - While this won't prevent you from starting the game,  the developer(s) of %s have found that "
+				+ "version %s of %s conflicts with their mod.\n"
+				+ "\t - It is heavily recommended to remove one of the mods.",
+				depCandidateVer, getCandidateName(candidate), depCandidateVer, getCandidateName(depCandidate));
 	}
 
-	private void appendBreakingError(StringBuilder errors, ModCandidate candidate, ModCandidate depCandidate) {
+	private void appendBreakingError(ModCandidate candidate, ModCandidate depCandidate, PrintWriter pw) {
 		final String depCandidateVer = getCandidateFriendlyVersion(depCandidate);
-		errors.append("but a matching version is present: ").append(depCandidate.getInfo().getVersion()).append("!");
-		errors.append("\n\t - The developer(s) of ").append(getCandidateName(candidate));
-		errors.append(" have found that version ").append(depCandidateVer).append(" of ").append(getCandidateName(depCandidate));
-		errors.append(" critically conflicts with their mod.");
-		errors.append("\n\t - You must remove one of the mods.");
+
+		pw.printf("but a matching version is present: %s!\n"
+				+ "\t - The developer(s) of %s have found that version %s of %s critically conflicts with their mod.\n"
+				+ "\t - You must remove one of the mods.",
+				depCandidate.getInfo().getVersion(), getCandidateName(candidate), depCandidateVer, getCandidateName(depCandidate));
 	}
 
-	private void appendJiJInfo(StringBuilder errors, Map<String, ModCandidate> result, ModCandidate candidate) {
+	private void appendJiJInfo(Map<String, ModCandidate> result, ModCandidate candidate, PrintWriter pw) {
 		if (candidate.getDepth() < 1) {
-			errors.append("\n\t - Mod ").append(getCandidateName(candidate))
-			.append(" v").append(getCandidateFriendlyVersion(candidate))
-			.append(" is being loaded from the user's mod directory.");
+			pw.printf("\n\t - Mod %s v%s is being loaded from the user's mod directory.",
+					getCandidateName(candidate), getCandidateFriendlyVersion(candidate));
 			return;
 		}
 
@@ -447,7 +413,7 @@ public class ModResolver {
 		URL sourceUrl = null;
 
 		try {
-			for (Map.Entry<String, List<Path>> entry : inMemoryCache.entrySet()) {
+			for (Map.Entry<String, List<Path>> entry : ModDiscoverer.inMemoryCache.entrySet()) {
 				for (Path path : entry.getValue()) {
 					URL url = UrlUtil.asUrl(path.normalize());
 
@@ -462,9 +428,8 @@ public class ModResolver {
 		}
 
 		if (sourceUrl == null) {
-			errors.append("\n\t - Mod ").append(getCandidateName(candidate))
-			.append(" v").append(getCandidateFriendlyVersion(candidate))
-			.append(" is being provided by <unknown mod>.");
+			pw.printf("\n\t - Mod %s v%s is being provided by <unknown mod>.",
+					getCandidateName(candidate), getCandidateFriendlyVersion(candidate));
 			return;
 		}
 
@@ -479,23 +444,18 @@ public class ModResolver {
 		}
 
 		if (srcCandidate == null) {
-			errors.append("\n\t - Mod ").append(getCandidateName(candidate))
-			.append(" v").append(getCandidateFriendlyVersion(candidate))
-			.append(" is being provided by <unknown mod: ")
-			.append(sourceUrl).append(">.");
+			pw.printf("\n\t - Mod %s v%s is being provided by <unknown mod: %s>.",
+					getCandidateName(candidate), getCandidateFriendlyVersion(candidate), sourceUrl);
 			return;
 		}
 
 		// now we have the proper data, yay
-		errors.append("\n\t - Mod ").append(getCandidateName(candidate))
-		.append(" v").append(getCandidateFriendlyVersion(candidate))
-		.append(" is being provided by ").append(getCandidateName(srcCandidate))
-		.append(" v").append(getCandidateFriendlyVersion(candidate))
-		.append('.');
+		pw.printf("\n\t - Mod %s v%s is being provided by %s v%s.",
+				getCandidateName(candidate), getCandidateFriendlyVersion(candidate), getCandidateName(srcCandidate), getCandidateFriendlyVersion(candidate));
 	}
 
 	private static String getCandidateName(ModCandidate candidate) {
-		return "'" + candidate.getInfo().getName() + "' (" + candidate.getInfo().getId() + ")";
+		return String.format("'%s' (%s)", candidate.getInfo().getName(), candidate.getInfo().getId());
 	}
 
 	private static String getCandidateFriendlyVersion(ModCandidate candidate) {
@@ -540,318 +500,5 @@ public class ModResolver {
 				return "unknown version"; // should be unreachable
 			}
 		}).collect(Collectors.joining(" or "));
-	}
-
-	/** @param errorList The list of errors. The returned list of errors all need to be prefixed with "it " in order to make sense. */
-	private static boolean isModIdValid(String modId, List<String> errorList) {
-		// A more useful error list for MOD_ID_PATTERN
-		if (modId.isEmpty()) {
-			errorList.add("is empty!");
-			return false;
-		}
-
-		if (modId.length() == 1) {
-			errorList.add("is only a single character! (It must be at least 2 characters long)!");
-		} else if (modId.length() > 64) {
-			errorList.add("has more than 64 characters!");
-		}
-
-		char first = modId.charAt(0);
-
-		if (first < 'a' || first > 'z') {
-			errorList.add("starts with an invalid character '" + first + "' (it must be a lowercase a-z - uppercase isn't allowed anywhere in the ID)");
-		}
-
-		Set<Character> invalidChars = null;
-
-		for (int i = 1; i < modId.length(); i++) {
-			char c = modId.charAt(i);
-
-			if (c == '-' || c == '_' || ('0' <= c && c <= '9') || ('a' <= c && c <= 'z')) {
-				continue;
-			}
-
-			if (invalidChars == null) {
-				invalidChars = new HashSet<>();
-			}
-
-			invalidChars.add(c);
-		}
-
-		if (invalidChars != null) {
-			StringBuilder error = new StringBuilder("contains invalid characters: ");
-			error.append(invalidChars.stream().map(value -> "'" + value + "'").collect(Collectors.joining(", ")));
-			errorList.add(error.append("!").toString());
-		}
-
-		assert errorList.isEmpty() == MOD_ID_PATTERN.matcher(modId).matches() : "Errors list " + errorList + " didn't match the mod ID pattern!";
-		return errorList.isEmpty();
-	}
-
-	@SuppressWarnings("serial")
-	static class UrlProcessAction extends RecursiveAction {
-		private final FabricLoaderImpl loader;
-		private final Map<String, ModCandidateSet> candidatesById;
-		private final URL url;
-		private final int depth;
-		private final boolean requiresRemap;
-
-		UrlProcessAction(FabricLoaderImpl loader, Map<String, ModCandidateSet> candidatesById, URL url, int depth, boolean requiresRemap) {
-			this.loader = loader;
-			this.candidatesById = candidatesById;
-			this.url = url;
-			this.depth = depth;
-			this.requiresRemap = requiresRemap;
-		}
-
-		@Override
-		protected void compute() {
-			FileSystemUtil.FileSystemDelegate jarFs;
-			Path path, modJson, rootDir;
-			URL normalizedUrl;
-
-			Log.debug(LogCategory.DISCOVERY, "Testing %s", url);
-
-			try {
-				path = UrlUtil.asPath(url).normalize();
-				// normalize URL (used as key for nested JAR lookup)
-				normalizedUrl = UrlUtil.asUrl(path);
-			} catch (UrlConversionException e) {
-				throw new RuntimeException("Failed to convert URL " + url + "!", e);
-			}
-
-			if (Files.isDirectory(path)) {
-				// Directory
-				modJson = path.resolve("fabric.mod.json");
-				rootDir = path;
-
-				if (loader.isDevelopmentEnvironment() && !Files.exists(modJson)) {
-					Log.warn(LogCategory.DISCOVERY, "Adding directory %s to mod classpath in development environment - workaround for Gradle splitting mods into two directories", path);
-					synchronized (launcherSyncObject) {
-						FabricLauncherBase.getLauncher().propose(url);
-					}
-				}
-			} else {
-				// JAR file
-				try {
-					jarFs = FileSystemUtil.getJarFileSystem(path, false);
-					modJson = jarFs.get().getPath("fabric.mod.json");
-					rootDir = jarFs.get().getRootDirectories().iterator().next();
-				} catch (IOException e) {
-					throw new RuntimeException("Failed to open mod JAR at " + path + "!");
-				} catch (ZipError e) {
-					throw new RuntimeException("Jar at " + path + " is corrupted, please redownload it!");
-				}
-			}
-
-			LoaderModMetadata[] info;
-
-			try {
-				info = new LoaderModMetadata[] { ModMetadataParser.parseMetadata(modJson) };
-			} catch (ParseMetadataException.MissingRequired e) {
-				throw new RuntimeException(String.format("Mod at \"%s\" has an invalid fabric.mod.json file! The mod is missing the following required field!", path), e);
-			} catch (MalformedJsonException | ParseMetadataException e) {
-				throw new RuntimeException(String.format("Mod at \"%s\" has an invalid fabric.mod.json file!", path), e);
-			} catch (NoSuchFileException e) {
-				Log.warn(LogCategory.DISCOVERY, "Non-Fabric mod JAR at \"%s\", ignoring", path);
-				info = new LoaderModMetadata[0];
-			} catch (IOException e) {
-				throw new RuntimeException(String.format("Failed to open fabric.mod.json for mod at \"%s\"!", path), e);
-			} catch (Throwable t) {
-				throw new RuntimeException(String.format("Failed to parse mod metadata for mod at \"%s\"", path), t);
-			}
-
-			for (LoaderModMetadata i : info) {
-				ModCandidate candidate = new ModCandidate(i, normalizedUrl, depth, requiresRemap);
-				boolean added;
-
-				if (candidate.getInfo().getId() == null || candidate.getInfo().getId().isEmpty()) {
-					throw new RuntimeException(String.format("Mod file `%s` has no id", candidate.getOriginUrl().getFile()));
-				}
-
-				if (!MOD_ID_PATTERN.matcher(candidate.getInfo().getId()).matches()) {
-					List<String> errorList = new ArrayList<>();
-					isModIdValid(candidate.getInfo().getId(), errorList);
-					StringBuilder fullError = new StringBuilder("Mod id `");
-					fullError.append(candidate.getInfo().getId()).append("` does not match the requirements because");
-
-					if (errorList.size() == 1) {
-						fullError.append(" it ").append(errorList.get(0));
-					} else {
-						fullError.append(":");
-
-						for (String error : errorList) {
-							fullError.append("\n  - It ").append(error);
-						}
-					}
-
-					throw new RuntimeException(fullError.toString());
-				}
-
-				for (String provides : candidate.getInfo().getProvides()) {
-					if (!MOD_ID_PATTERN.matcher(provides).matches()) {
-						List<String> errorList = new ArrayList<>();
-						isModIdValid(provides, errorList);
-						StringBuilder fullError = new StringBuilder("Mod id provides `");
-						fullError.append(provides).append("` does not match the requirements because");
-
-						if (errorList.size() == 1) {
-							fullError.append(" it ").append(errorList.get(0));
-						} else {
-							fullError.append(":");
-
-							for (String error : errorList) {
-								fullError.append("\n  - It ").append(error);
-							}
-						}
-
-						throw new RuntimeException(fullError.toString());
-					}
-				}
-
-				added = candidatesById.computeIfAbsent(candidate.getInfo().getId(), ModCandidateSet::new).add(candidate);
-
-				if (!added) {
-					Log.debug(LogCategory.DISCOVERY, "%s already present as %s", candidate.getOriginUrl(), candidate);
-				} else {
-					Log.debug(LogCategory.DISCOVERY, "Adding %s as %s", candidate.getOriginUrl(), candidate);
-
-					List<Path> jarInJars = inMemoryCache.computeIfAbsent(candidate.getOriginUrl().toString(), (u) -> {
-						Log.debug(LogCategory.DISCOVERY, "Searching for nested JARs in %s", candidate);
-						Log.debug(LogCategory.DISCOVERY, u);
-						Collection<NestedJarEntry> jars = candidate.getInfo().getJars();
-						List<Path> list = new ArrayList<>(jars.size());
-
-						for (NestedJarEntry jar : jars) {
-							Path modPath = rootDir.resolve(jar.getFile().replace("/", rootDir.getFileSystem().getSeparator()));
-
-							if (!Files.isDirectory(modPath) && modPath.toString().endsWith(".jar")) {
-								// TODO: pre-check the JAR before loading it, if possible
-								Log.debug(LogCategory.DISCOVERY, "Found nested JAR: %s", modPath);
-								Path dest = inMemoryFs.getPath(UUID.randomUUID() + ".jar");
-
-								try {
-									Files.copy(modPath, dest);
-								} catch (IOException e) {
-									throw new RuntimeException("Failed to load nested JAR " + modPath + " into memory (" + dest + ")!", e);
-								}
-
-								list.add(dest);
-							}
-						}
-
-						return list;
-					});
-
-					if (!jarInJars.isEmpty()) {
-						invokeAll(
-								jarInJars.stream()
-								.map((p) -> {
-									try {
-										return new UrlProcessAction(loader, candidatesById, UrlUtil.asUrl(p.normalize()), depth + 1, requiresRemap);
-									} catch (UrlConversionException e) {
-										throw new RuntimeException("Failed to turn path '" + p.normalize() + "' into URL!", e);
-									}
-								}).collect(Collectors.toList()));
-					}
-				}
-			}
-
-			/* if (jarFs != null) {
-				jarFs.close();
-			} */
-		}
-	}
-
-	public Map<String, ModCandidate> resolve(FabricLoaderImpl loader) throws ModResolutionException {
-		ConcurrentMap<String, ModCandidateSet> candidatesById = new ConcurrentHashMap<>();
-
-		long time1 = System.currentTimeMillis();
-		Queue<UrlProcessAction> allActions = new ConcurrentLinkedQueue<>();
-		ForkJoinPool pool = new ForkJoinPool(Math.max(1, Runtime.getRuntime().availableProcessors() - 1));
-
-		for (ModCandidateFinder f : candidateFinders) {
-			f.findCandidates(loader, (u, requiresRemap) -> {
-				UrlProcessAction action = new UrlProcessAction(loader, candidatesById, u, 0, requiresRemap);
-				allActions.add(action);
-				pool.execute(action);
-			});
-		}
-
-		// add builtin mods
-		for (BuiltinMod mod : loader.getGameProvider().getBuiltinMods()) {
-			addBuiltinMod(candidatesById, mod);
-		}
-
-		// Add the current Java version
-		try {
-			addBuiltinMod(candidatesById, new BuiltinMod(
-					new File(System.getProperty("java.home")).toURI().toURL(),
-					new BuiltinModMetadata.Builder("java", System.getProperty("java.specification.version").replaceFirst("^1\\.", ""))
-					.setName(System.getProperty("java.vm.name"))
-					.build()));
-		} catch (MalformedURLException e) {
-			throw new ModResolutionException("Could not add Java to the dependency constraints", e);
-		}
-
-		boolean tookTooLong = false;
-		Throwable exception = null;
-
-		try {
-			pool.shutdown();
-			// Comment out for debugging
-			pool.awaitTermination(30, TimeUnit.SECONDS);
-
-			for (UrlProcessAction action : allActions) {
-				if (!action.isDone()) {
-					tookTooLong = true;
-				} else {
-					Throwable t = action.getException();
-
-					if (t != null) {
-						if (exception == null) {
-							exception = t;
-						} else {
-							exception.addSuppressed(t);
-						}
-					}
-				}
-			}
-		} catch (InterruptedException e) {
-			throw new ModResolutionException("Mod resolution took too long!", e);
-		}
-
-		if (tookTooLong) {
-			throw new ModResolutionException("Mod resolution took too long!");
-		}
-
-		if (exception != null) {
-			throw new ModResolutionException("Mod resolution failed!", exception);
-		}
-
-		long time2 = System.currentTimeMillis();
-		Map<String, ModCandidate> result = findCompatibleSet(candidatesById);
-
-		long time3 = System.currentTimeMillis();
-		Log.debug(LogCategory.RESOLUTION, "Mod resolution detection time: " + (time2 - time1) + "ms");
-		Log.debug(LogCategory.RESOLUTION, "Mod resolution time: " + (time3 - time2) + "ms");
-
-		for (ModCandidate candidate : result.values()) {
-			if (candidate.getInfo().getSchemaVersion() < ModMetadataParser.LATEST_VERSION) {
-				Log.warn(LogCategory.METADATA, "Mod ID %s uses outdated schema version: %d < %d", candidate.getInfo().getId(), candidate.getInfo().getSchemaVersion(), ModMetadataParser.LATEST_VERSION);
-			}
-
-			candidate.getInfo().emitFormatWarnings();
-		}
-
-		return result;
-	}
-
-	private void addBuiltinMod(ConcurrentMap<String, ModCandidateSet> candidatesById, BuiltinMod mod) {
-		candidatesById.computeIfAbsent(mod.metadata.getId(), ModCandidateSet::new).add(new ModCandidate(new BuiltinMetadataWrapper(mod.metadata), mod.url, 0, false));
-	}
-
-	public static FileSystem getInMemoryFs() {
-		return inMemoryFs;
 	}
 }
